Extract App trait into ctaphid-app
Applications don’t need to depend on the dispatch implementation.  They
only need the App trait.  This patch moves the trait into a separate
crate.
