+++ Cargo.toml
@@ -1,32 +1,17 @@
-[package]
-name = "ctaphid-dispatch"
-version = "0.1.1"
-authors = ["Conor Patrick <conor@solokeys.com>", "Nicolas Stalder <n@stalder.io>"]
+[workspace]
+members = ["app", "dispatch"]
+resolver = "2"
+
+[workspace.package]
+authors = ["Conor Patrick <conor@solokeys.com>", "Nicolas Stalder <n@stalder.io>", "The Trussed developers"]
 edition = "2021"
 license = "Apache-2.0 OR MIT"
-description = "Dispatch layer after usbd-ctaphid"
 repository = "https://github.com/solokeys/ctaphid-dispatch"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-delog = "0.1"
+[workspace.dependencies]
 heapless = "0.7"
-heapless-bytes = "0.3"
-interchange = "0.3.0"
-ref-swap = "0.1.2"
 trussed-core = "0.1"
 
-[features]
-default = []
-std = ["delog/std"]
-
-log-all = []
-log-none = []
-log-info = []
-log-debug = []
-log-warn = []
-log-error = []
-
 [patch.crates-io]
+ctaphid-app.path = "app"
 trussed-core = { git = "https://github.com/trussed-dev/trussed.git", rev = "8a9bb9d79e40c2dcf026cf611302109f776bec83" }
