+++ app/CHANGELOG.md
@@ -1,0 +1,5 @@
+# Changelog
+
+## Unreleased
+
+- Extract `app` and `command` modules from `ctaphid-dispatch` 0.1 into a separate crate.
